Move option methods to a new class
